refactor: rename alchemy to domain name
