+++ src/repositories/api/nft.endpoint.spec.ts
@@ -1,11 +1,11 @@
-import { alchemyNftMetadataMock } from './mocks/mock'
+import { nftMetadataMock } from './mocks/mock'
 import { createNftEndpoint } from './nft.endpoint'
 
 describe('NftEndpoint', () => {
   const { getNftMetadata, clientExposedForTestingOnly } = createNftEndpoint()
 
-  it('fetches data from alchemy', async () => {
-    vi.spyOn(clientExposedForTestingOnly.nft, 'getNftMetadata').mockResolvedValue(alchemyNftMetadataMock)
+  it('fetches data', async () => {
+    vi.spyOn(clientExposedForTestingOnly.nft, 'getNftMetadata').mockResolvedValue(nftMetadataMock)
     const result = await getNftMetadata('aValidAddress', '1')
 
     expect(result.address).toEqual('aValidAddress')
